requirements.txt
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
faker==37.11.0
filelock==3.20.0
fsspec==2025.9.0
groq==0.32.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
litestar==2.18.0
litestar-htmx==0.5.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.7.0
multipart==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
ollama==0.5.2
packaging==25.0
pgvector==0.4.1
pillow==11.3.0
polyfactory==2.22.2
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
pygments==2.19.2
python-dotenv==1.0.0
python-telegram-bot==22.5
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-click==1.9.2
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
sniffio==1.3.1
sqlalchemy==2.0.23
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
